MOVE LIBRARY INTERNALS INTO THE INTEROP NAMESPACE
Move RuntimeShim and SodiumLibrary classes into the Sodium.Interop
namespace in order to clearly seperate them from the puplic API.

Add SodiumRuntimeConfig which holds the configurable bits of the
library. This class currently doesn't contain much and is more or less a
placeholder for the shadow copying workaround where the user will need
to provide some runtime config hints.
