+++ libsodium-net/KeyPair.cs
@@ -1,11 +1,12 @@
 using System;
 using System.Security.Cryptography;
 using Sodium.Exceptions;
+using Sodium.Interop;
 
 namespace Sodium
 {
   /// <summary>A public / private key pair.</summary>
-  public partial class KeyPair : IDisposable
+  public class KeyPair : IDisposable
   {
     private readonly byte[] _publicKey;
     private readonly byte[] _privateKey;
@@ -23,7 +24,7 @@ namespace Sodium
       _publicKey = publicKey;
 
       _privateKey = privateKey;
-      _ProtectKey();
+      RuntimeShim.ProtectMemory(_privateKey);
     }
 
     ~KeyPair()
@@ -42,10 +43,10 @@ namespace Sodium
     {
       get
       {
-        _UnprotectKey();
+        RuntimeShim.UnprotectMemory(_privateKey);
         var tmp = new byte[_privateKey.Length];
         Array.Copy(_privateKey, tmp, tmp.Length);
-        _ProtectKey();
+        RuntimeShim.ProtectMemory(_privateKey);
 
         return tmp;
       }
@@ -57,25 +58,5 @@ namespace Sodium
       if (_privateKey != null && _privateKey.Length > 0)
         Array.Clear(_privateKey, 0, _privateKey.Length);
     }
-
-    partial void _ProtectKey();
-    partial void _UnprotectKey();
   }
-
-#if NET46
-  public partial class KeyPair
-  {
-    partial void _ProtectKey()
-    {
-      if (!SodiumLibrary.IsRunningOnMono)
-        ProtectedMemory.Protect(_privateKey, MemoryProtectionScope.SameProcess);
-    }
-
-    partial void _UnprotectKey()
-    {
-      if (!SodiumLibrary.IsRunningOnMono)
-        ProtectedMemory.Unprotect(_privateKey, MemoryProtectionScope.SameProcess);
-    }
-  }
-#endif
 }
