+++ libsodium-net/OneTimeAuth.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text;
 using Sodium.Exceptions;
+using Sodium.Interop;
 
 namespace Sodium
 {
