+++ libsodium-net/ScalarMult.cs
@@ -1,4 +1,5 @@
 using Sodium.Exceptions;
+using Sodium.Interop;
 
 namespace Sodium
 {
