+++ libsodium-net/SecretAeadAes.cs
@@ -2,6 +2,7 @@ using System;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
 using Sodium.Exceptions;
+using Sodium.Interop;
 
 namespace Sodium
 {
@@ -85,7 +86,7 @@ namespace Sodium
 
       //remove the trailing nulls from the array
       var tmp = new byte[cipherLength];
-      SodiumLibrary.Copy(cipher, 0, tmp, 0, cipherLength);
+      RuntimeShim.Copy(cipher, 0, tmp, 0, cipherLength);
 
       return tmp;
     }
@@ -141,7 +142,7 @@ namespace Sodium
 
       //remove the trailing nulls from the array
       var tmp = new byte[messageLength];
-      SodiumLibrary.Copy(message, 0, tmp, 0, messageLength);
+      RuntimeShim.Copy(message, 0, tmp, 0, messageLength);
 
       return tmp;
     }
