+++ libsodium-net/SecretBox.cs
@@ -2,6 +2,7 @@
 using System.Security.Cryptography;
 using System.Text;
 using Sodium.Exceptions;
+using Sodium.Interop;
 
 namespace Sodium
 {
