+++ libsodium-net/SodiumCore.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Sodium.Interop;
+
 namespace Sodium
 {
   /// <summary>
