+++ libsodium-net/StreamEncryption.cs
@@ -1,6 +1,7 @@
 ﻿using System.Security.Cryptography;
 using System.Text;
 using Sodium.Exceptions;
+using Sodium.Interop;
 
 namespace Sodium
 {
