+++ libsodium-net/Utilities.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using System.Text;
+using Sodium.Interop;
 
 namespace Sodium
 {
